Add possibility to define average_length for monsters in JSON.
This change will allow to define length for monsters in JSON. It complements Kevins
fix for issue #40446. Along with that it also implements check for quartered corpse.

references #40446

This commit adds definition of "average_length" from JSON and adds it as
attribute longest_side of the monster type object. Referrence is then used
to define "length" in the item generation. Utilizing Kevins' addition
to default based on the items volume.